requirements.txt
gunicorn==19.9.0
django-heroku==0.3.1
metar==1.8.0
requests==2.21.0
pandas==0.23.3
django-bmemcached==0.3.0
